stomp-client/requirements.txt
docopt==0.6.2
stomp.py==4.1.19
